refactor: Move completion package one level up
