+++ cmd/monaco/deploy/command.go
@@ -19,7 +19,7 @@ package deploy
 import (
 	"fmt"
 	"github.com/dynatrace/dynatrace-configuration-as-code/cmd/monaco/cmdutils"
-	"github.com/dynatrace/dynatrace-configuration-as-code/cmd/monaco/runner/completion"
+	"github.com/dynatrace/dynatrace-configuration-as-code/cmd/monaco/completion"
 	"github.com/dynatrace/dynatrace-configuration-as-code/internal/files"
 	"github.com/dynatrace/dynatrace-configuration-as-code/internal/log"
 	"github.com/spf13/afero"
